+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -12,6 +12,7 @@
 - PhoneNumbersModelFactory static method AcquiredPhoneNumber is renamed to PurchasedPhoneNumber.
 - PurchasePhoneNumbersOperation extends Operation instead of Operation<Response>.
 - ReleasePhoneNumberOperation extends Operation instead of Operation<Response>.
+- Removed PhoneNumberOperationStatus and PhoneNumberOperationType.
 
 ## 1.0.0-beta.5 (2021-03-09)