+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -10,8 +10,8 @@
   - GetPhoneNumbers -> GetPurchasedPhoneNumbers.
   - GetPhoneNumbersAsync -> GetPurchasedPhoneNumbersAsync.
 - PhoneNumbersModelFactory static method AcquiredPhoneNumber is renamed to PurchasedPhoneNumber.
-- PurchasePhoneNumbersOperation extends Operation instead of Operation<Response>.
-- ReleasePhoneNumberOperation extends Operation instead of Operation<Response>.
+- PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation extend Operation instead of Operation<Response>.
+- Added protected constructor to PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation for mocking.
 - Removed PhoneNumberOperationStatus and PhoneNumberOperationType.
 
 ## 1.0.0-beta.5 (2021-03-09)