+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.6 (Unreleased)
 
+### Added
+- Added protected constructor to PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation for mocking.
+
 ### Breaking Changes
 - AcquiredPhoneNumber class is renamed to PurchasedPhoneNumber.
 - PhoneNumbersClient methods renamed:
@@ -11,7 +14,6 @@
   - GetPhoneNumbersAsync -> GetPurchasedPhoneNumbersAsync.
 - PhoneNumbersModelFactory static method AcquiredPhoneNumber is renamed to PurchasedPhoneNumber.
 - PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation extend Operation instead of Operation<Response>.
-- Added protected constructor to PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation for mocking.
 - Removed PhoneNumberOperationStatus and PhoneNumberOperationType.
 
 ## 1.0.0-beta.5 (2021-03-09)