+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.6 (Unreleased)
+## 1.0.0-beta.6 (2021-03-29)
 
 ### Added
 - Added protected constructor to PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation for mocking.